core: Add basic root object type
This is super basic right now, but will be extended to support a trivial
refcount based underlying object for all of our types. This will allow us
to blindly instaniate objects and allow the attached objects to care about
their lifecycle, as well as allow attaching to multiple scenes, etc.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>