core: Implement a very very basic camera. (fixes #6)
This is kinda buggy and inefficient but we'll use this for now to
implement the next parts of the engine we need. Long story short our
camera can be clipped to a world bound, and we have a view within
those bounds for display.

If an entity falls outside the camera viewport, its no longer rendered.
If its inside, its rendered at a given offset.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>