core: Start respecting camera bounds and draw at offset (#13)
We still need to cull how many rows and columns we draw, as a TileSheet
is always considered to be at the root (0, 0) position in the renderable
universe.

In future we'll need to see how far offset the camera is to our tilesheet
and render only the visible rows and columns. To that purpose, we link the
math library.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>